batch: Move line matching into package
The line-matching package owns mapping and workspace storage while the
algorithm that uses both remains at the batch package root.

That split prevents the package from containing the operation its
lower-level values exist to support.

This commit moves the matching algorithm beside its mapping and workspace
dependencies while retaining the selected import path for existing tests.

Later commits will add comparison, sequence, range, and lineage helpers to
the same bounded package.
